refactor: remove node-bridge
