--- packages/node-bridge/.gitignore
@@ -1,2 +1,0 @@
-/helpers.js
-/source-map-support.js
