--- packages/node-bridge/index.js
@@ -1,3 +1,0 @@
-const path = require('path');
-
-module.exports = path.join(__dirname, 'bridge.js');
