--- packages/node-bridge/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@vercel/node-bridge",
-  "version": "4.0.1",
-  "license": "Apache-2.0",
-  "main": "./index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/vercel/vercel.git",
-    "directory": "packages/node-bridge"
-  },
-  "files": [
-    "bridge.*",
-    "launcher.*",
-    "index.js",
-    "helpers.js",
-    "source-map-support.js"
-  ],
-  "scripts": {
-    "build": "node build.js",
-    "test": "jest --env node --verbose --runInBand --bail",
-    "test-unit": "pnpm test"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "8.10.19",
-    "@types/node": "14.18.33",
-    "content-type": "1.0.4",
-    "cookie": "0.4.0",
-    "etag": "1.8.1",
-    "execa": "3.2.0",
-    "fs-extra": "10.0.0",
-    "jsonlines": "0.1.1",
-    "test-listen": "1.1.0",
-    "typescript": "4.3.4"
-  }
-}
