--- packages/node-bridge/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "checkJs": true,
-    "lib": ["ES2020", "DOM"],
-    "noEmit": true,
-    "noImplicitReturns": true,
-    "strict": true,
-    "target": "ES2020",
-    "declaration": true,
-    "module": "commonjs",
-    "skipLibCheck": true
-  },
-  "include": ["helpers.ts", "bridge.js", "launcher.js"],
-  "exclude": ["node_modules"]
-}
