--- packages/node-bridge/turbo.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://turborepo.org/schema.json",
-  "extends": ["//"],
-  "pipeline": {
-    "build": {
-      "outputs": ["helpers.js", "source-map-support.js"]
-    }
-  }
-}
