+++ packages/node/src/edge-functions/edge-handler.ts
@@ -1,17 +1,17 @@
-import { IncomingMessage } from 'http';
-import { VercelProxyResponse } from '@vercel/node-bridge/types';
-import exitHook from 'exit-hook';
-import { EdgeRuntime, runServer } from 'edge-runtime';
 import type { EdgeContext } from '@edge-runtime/vm';
-import esbuild from 'esbuild';
-import { fetch } from 'undici';
+import type { VercelProxyResponse } from '../types';
 import { createEdgeWasmPlugin, WasmAssets } from './edge-wasm-plugin';
-import { serializeRequest, entrypointToOutputPath, logError } from '../utils';
-import { readFileSync } from 'fs';
 import {
   createNodeCompatPlugin,
   NodeCompatBindings,
 } from './edge-node-compat-plugin';
+import { EdgeRuntime, runServer } from 'edge-runtime';
+import { fetch } from 'undici';
+import { IncomingMessage } from 'http';
+import { readFileSync } from 'fs';
+import { serializeRequest, entrypointToOutputPath, logError } from '../utils';
+import esbuild from 'esbuild';
+import exitHook from 'exit-hook';
 
 const NODE_VERSION_MAJOR = process.version.match(/^v(\d+)\.\d+/)?.[1];
 const NODE_VERSION_IDENTIFIER = `node${NODE_VERSION_MAJOR}`;
